# src/generated/openapi.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Employee": {
				"description": "Employee model from BHXH API (code 067)\nNote: Uses string-based id to match API response format",
				"properties": {
					"id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						],
						"description": "Employee ID (as string from API)"
					},
					"Hoten": {
						"type": "string",
						"description": "Full name"
					},
					"Masobhxh": {
						"type": "string",
						"description": "BHXH number"
					},
					"NgaySinh": {
						"type": "string",
						"description": "Birth date"
					},
					"GioiTinh": {
						"type": "string",
						"description": "Gender"
					},
					"DiaChi": {
						"type": "string",
						"description": "Address"
					},
					"maPhongBan": {
						"type": "string",
						"description": "Department code"
					},
					"maTinhTrang": {
						"type": "string",
						"description": "Status code"
					},
					"NgayBatDau": {
						"type": "string",
						"description": "Work start date"
					},
					"chucVu": {
						"type": "string",
						"description": "Position"
					},
					"mucLuong": {
						"type": "string",
						"description": "Salary"
					},
					"soDienThoai": {
						"type": "string",
						"description": "Phone"
					},
					"email": {
						"type": "string",
						"description": "Email"
					},
					"ghiChu": {
						"type": "string",
						"description": "Notes"
					}
				},
				"type": "object",
				"additionalProperties": {}
			},
			"EmployeeListResponse": {
				"description": "Employee list response with pagination",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"timing": {
						"properties": {
							"totalMs": {
								"type": "number",
								"format": "double"
							},
							"fetchMs": {
								"type": "number",
								"format": "double"
							},
							"sessionMs": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalMs",
							"fetchMs",
							"sessionMs"
						],
						"type": "object",
						"description": "Timing metrics"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Employee"
						},
						"type": "array",
						"description": "Array of employee records"
					},
					"meta": {
						"properties": {
							"unit": {
								"type": "string"
							},
							"count": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"unit",
							"count",
							"total"
						],
						"type": "object",
						"description": "Response metadata"
					}
				},
				"required": [
					"success",
					"timing",
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmployeeDetailResponse": {
				"description": "Employee detail response (Code 172)",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/Employee"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FullEmployeeDetailsResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/FullEmployeeDetails"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmployeeBulkUploadResponse": {
				"description": "Employee bulk upload response (Code 112)",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"processed": {
						"type": "number",
						"format": "double",
						"description": "Number of employees processed"
					},
					"errors": {
						"items": {
							"properties": {
								"message": {
									"type": "string"
								},
								"row": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"message",
								"row"
							],
							"type": "object"
						},
						"type": "array",
						"description": "Upload errors if any"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmployeeUpdateResponse": {
				"description": "Employee update response (Code 068)",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/Employee"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmployeeUpdateRequest": {
				"description": "Employee update request (Code 068)\nFull employee data structure for update operation",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "Employee database ID (required for update)"
					},
					"soHopDong": {
						"type": "string",
						"nullable": true,
						"description": "Contract number"
					},
					"loaiHopDong": {
						"type": "string",
						"nullable": true,
						"description": "Contract type"
					},
					"ngayHieuLuc": {
						"type": "string",
						"nullable": true,
						"description": "Effective date"
					},
					"ngayKy": {
						"type": "string",
						"nullable": true,
						"description": "Signing date"
					},
					"heSoLuong": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "Salary coefficient"
					},
					"mucLuong": {
						"type": "string",
						"description": "Salary amount"
					},
					"phuCapCV": {
						"type": "string",
						"description": "Position allowance"
					},
					"phuCapTNNghe": {
						"type": "string",
						"nullable": true,
						"description": "Seniority allowance"
					},
					"phuCapTNVK": {
						"type": "string",
						"nullable": true,
						"description": "Dangerous job allowance"
					},
					"phuCapLuong": {
						"type": "string",
						"description": "Salary allowance"
					},
					"phuCapBoSung": {
						"type": "string",
						"description": "Additional allowance"
					},
					"phuCapKhac": {
						"type": "string",
						"description": "Other allowance"
					},
					"noiDKKCB": {
						"type": "string",
						"description": "Healthcare facility"
					},
					"Matinhbenhvien": {
						"type": "string",
						"description": "Province code for hospital"
					},
					"Mabenhvien": {
						"type": "string",
						"description": "Hospital code"
					},
					"ghiChu": {
						"type": "string",
						"description": "Notes"
					},
					"listThanhVien": {
						"items": {
							"$ref": "#/components/schemas/EmployeeFamilyMember"
						},
						"type": "array",
						"description": "Family members list"
					},
					"hoTenChuHo": {
						"type": "string",
						"nullable": true,
						"description": "Household name"
					},
					"maSo_HoGiaDinh": {
						"type": "string",
						"nullable": true,
						"description": "Household ID"
					},
					"dienThoaiChuHo": {
						"type": "string",
						"nullable": true,
						"description": "Household phone"
					},
					"diaChi_HK": {
						"type": "string",
						"description": "Permanent residence address"
					},
					"tinh_hokhau": {
						"type": "string",
						"description": "Province for household registration"
					},
					"huyen_hokhau": {
						"type": "string",
						"description": "District for household registration"
					},
					"xa_hokhau": {
						"type": "string",
						"description": "Ward for household registration"
					},
					"loaiGiayTo": {
						"type": "string",
						"nullable": true,
						"description": "Document type"
					},
					"soGiayTo": {
						"type": "string",
						"nullable": true,
						"description": "Document number"
					},
					"diaChiGiaDinh": {
						"type": "string",
						"description": "Family address"
					},
					"maTinhGiaDinh": {
						"type": "string",
						"description": "Province for family"
					},
					"maHuyenGiaDinh": {
						"type": "string",
						"description": "District for family"
					},
					"maXaGiaDinh": {
						"type": "string",
						"description": "Ward for family"
					},
					"Tentinh": {
						"type": "string",
						"nullable": true,
						"description": "Province name"
					},
					"Tenhuyen": {
						"type": "string",
						"nullable": true,
						"description": "District name"
					},
					"Tenxa": {
						"type": "string",
						"nullable": true,
						"description": "Ward name"
					},
					"diaChi_HK_full": {
						"type": "string",
						"description": "Full household address"
					},
					"diaChi_ks": {
						"type": "string",
						"description": "Current residence address"
					},
					"Hoten": {
						"type": "string",
						"description": "Full name"
					},
					"tinhTrang": {
						"type": "string",
						"description": "Employment status"
					},
					"Ngaysinh": {
						"type": "string",
						"description": "Birth date"
					},
					"email": {
						"type": "string",
						"nullable": true,
						"description": "Email"
					},
					"maSoThue": {
						"type": "string",
						"nullable": true,
						"description": "Tax ID"
					},
					"maPhongBan": {
						"type": "string",
						"description": "Department code"
					},
					"Masobhxh": {
						"type": "string",
						"description": "BHXH number"
					},
					"maNLD": {
						"type": "string",
						"nullable": true,
						"description": "Employee code"
					},
					"quocTich": {
						"type": "string",
						"description": "Nationality"
					},
					"Gioitinh": {
						"type": "string",
						"description": "Gender (0=female, 1=male)"
					},
					"soDienThoai": {
						"type": "string",
						"nullable": true,
						"description": "Phone number"
					},
					"danToc": {
						"type": "string",
						"description": "Ethnicity"
					},
					"taiKhoan": {
						"type": "string",
						"nullable": true,
						"description": "Bank account"
					},
					"chucVu": {
						"type": "string",
						"description": "Position"
					},
					"maSoBHYT": {
						"type": "string",
						"nullable": true,
						"description": "Health insurance number"
					},
					"nganHang": {
						"type": "string",
						"nullable": true,
						"description": "Bank name"
					},
					"Manganhang": {
						"type": "string",
						"nullable": true,
						"description": "Bank code"
					},
					"Matinhnganhang": {
						"type": "string",
						"nullable": true,
						"description": "Bank province code"
					},
					"soCMND": {
						"type": "string",
						"description": "ID card number"
					},
					"noiCapCMND": {
						"type": "string",
						"nullable": true,
						"description": "ID card issuance place"
					},
					"ngayCapCMND": {
						"type": "string",
						"nullable": true,
						"description": "ID card issuance date"
					},
					"noiCapKS": {
						"type": "string",
						"description": "ID card issuance place (detailed)"
					},
					"maTinhKS": {
						"type": "string",
						"description": "Province for ID card issuance"
					},
					"maHuyenKS": {
						"type": "string",
						"nullable": true,
						"description": "District for ID card issuance"
					},
					"maXaKS": {
						"type": "string",
						"description": "Ward for ID card issuance"
					},
					"nghiViec": {
						"type": "boolean",
						"description": "Resigned flag"
					},
					"diaChiNN": {
						"type": "string",
						"description": "Permanent address"
					},
					"maTinhNN": {
						"type": "string",
						"description": "Province for permanent address"
					},
					"maHuyenNN": {
						"type": "string",
						"nullable": true,
						"description": "District for permanent address"
					},
					"maXaNN": {
						"type": "string",
						"description": "Ward for permanent address"
					},
					"dcCuTheNN": {
						"type": "string",
						"description": "Detailed permanent address"
					},
					"Ccns": {
						"type": "string",
						"description": "C2NS status"
					},
					"maHoGiaDinh": {
						"type": "string",
						"description": "Household code"
					},
					"maVungSS": {
						"type": "string",
						"nullable": true,
						"description": "Region code"
					},
					"NgaysinhPicker": {
						"type": "string",
						"description": "Birth date (picker format)"
					},
					"diaChiCuThe_dangSS": {
						"type": "string",
						"description": "Detailed current address"
					},
					"maTinh_dangSS": {
						"type": "string",
						"description": "Province for current address"
					},
					"maHuyen_dangSS": {
						"type": "string",
						"nullable": true,
						"description": "District for current address"
					},
					"maXa_dangSS": {
						"type": "string",
						"description": "Ward for current address"
					},
					"diaChi_dangSS": {
						"type": "string",
						"description": "Current address"
					},
					"noiLamViec": {
						"type": "string",
						"nullable": true,
						"description": "Workplace"
					},
					"masobhxhuser": {
						"type": "string",
						"description": "BHXH username (for API call)"
					},
					"macoquanuser": {
						"type": "string",
						"description": "Agency code (for API call)"
					},
					"loaidoituonguser": {
						"type": "string",
						"description": "Object type (for API call)"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EmployeeSyncResponse": {
				"description": "Employee sync response (Code 156)\nOfficial employee data from central BHXH system",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/EmployeeOfficialData"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "X-API-Key",
				"description": "API key for authentication",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "bhxh-api",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
...
		"/api/v1/employees": {
			"get": {
				"operationId": "GetEmployees",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmployeeListResponse"
								}
							}
						}
					},
					"500": {
						"description": "Failed to fetch employees",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get all employees for current unit",
				"security": [],
				"parameters": [
					{
						"description": "Filter by employee code",
						"in": "query",
						"name": "maNguoiLaoDong",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by name",
						"in": "query",
						"name": "ten",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by department",
						"in": "query",
						"name": "maPhongBan",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by status",
						"in": "query",
						"name": "maTinhTrang",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by BHXH number",
						"in": "query",
						"name": "MaSoBhxh",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Page number (default 1)",
						"in": "query",
						"name": "PageIndex",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Page size (default 100)",
						"in": "query",
						"name": "PageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/employees/{employeeId}": {
			"get": {
				"operationId": "GetEmployeeById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmployeeDetailResponse"
								}
							}
						}
					},
					"404": {
						"description": "Employee not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Failed to fetch employee",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get employee by ID (Code 172)\nFetches employee detail from internal record by ID",
				"security": [],
				"parameters": [
					{
						"description": "Employee ID (internal record ID)",
						"in": "path",
						"name": "employeeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateEmployee",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmployeeUpdateResponse"
								}
							}
						}
					},
					"404": {
						"description": "Employee not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Failed to update employee",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Update employee (Code 068)\nNote: Requires full employee data structure - use GET /employees/{id} first to retrieve current data",
				"security": [],
				"parameters": [
					{
						"description": "Employee ID",
						"in": "path",
						"name": "employeeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Full employee update data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmployeeUpdateRequest",
								"description": "Full employee update data"
							}
						}
					}
				}
			}
		},
		"/api/v1/employees/{employeeId}/details": {
			"get": {
				"operationId": "GetFullEmployeeDetails",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FullEmployeeDetailsResponse"
								}
							}
						}
					},
					"404": {
						"description": "Employee not found",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Failed to fetch employee details",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get full employee details by ID (Code 117)\nFetches comprehensive employee data including contracts, salary, family, history",
				"security": [],
				"parameters": [
					{
						"description": "Employee ID (internal record ID)",
						"in": "path",
						"name": "employeeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/employees/upload": {
			"post": {
				"operationId": "UploadEmployees",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmployeeBulkUploadResponse"
								}
							}
						}
					},
					"500": {
						"description": "Failed to upload employees",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Bulk upload employees from Excel file (Code 112)\nNote: Requires multipart/form-data with Excel file",
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/employees/{employeeId}/sync": {
			"get": {
				"operationId": "SyncEmployeeById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EmployeeSyncResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid sync parameters",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Failed to sync employee",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"error": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"error"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Sync employee with central BHXH system (Code 156)\nFetches official employee data from central BHXH using Social Security Number",
				"security": [],
				"parameters": [
					{
						"description": "Employee ID (for route consistency)",
						"in": "path",
						"name": "employeeId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "10-digit Social Security Number",
						"in": "query",
						"name": "masoBhxh",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Social Security Agency code",
						"in": "query",
						"name": "maCqbh",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}